Basic working support for connecting to Keycloak.
This code is essentially the VKProvider with the GET to the userInfo
url replaced with a form data POST that Keycloak seems to support.

Whilst Keycloak supports OIDC, this is really just using OAuth2 under
the hood. OpenId (xml based) != OpenIdConnect (json based) and I don't
see any Silhouette providers with this support yet.

There are no tests for this module yet.

This module relies on the caller setting apiURL to the userInfo
endpoint. apiURL is an Option[String] and throws an Exception when
it's None, which is not ideal.

Finally the LoginInfo gets set to "keycloak" and the user's email
(throwing an exception if it's not - I wish these functions would return Eithers!).
It should probably be setting that to preferred_username and we should
probably find a way to allow the mapping of all the other mappable profile fields.

Keycloak supports a collection of login mechanisms, OIDC, OAuth2, OAuth1. This
provider may not be suitable for other people's older or more esoteric or established
Keycloak services which are using slightly older protocols.